+++ src/main/java/com/kakaotech/team14backend/inner/post/model/PostRandomFetcher.java
@@ -1,7 +1,6 @@
 package com.kakaotech.team14backend.inner.post.model;
 
 import com.kakaotech.team14backend.exception.PostLevelOutOfRangeException;
-import com.kakaotech.team14backend.outer.post.dto.RandomIndexes;
 import org.springframework.stereotype.Component;
 
 import java.util.HashMap;
