+++ src/test/java/com/kakaotech/team14backend/inner/post/model/PostRandomFetcherTest.java
@@ -1,6 +1,5 @@
 package com.kakaotech.team14backend.inner.post.model;
 
-import com.kakaotech.team14backend.outer.post.dto.RandomIndexes;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
